+++ waveforms/misc_waveforms.odin
@@ -3,9 +3,9 @@ package waveforms
 import "core:math"
 import "../mathx"
 
-explicit_harmonics :: proc(harmonics: []f64, volume_limiter: f64, phase: f64) -> f64 {
-    val: f64
-    i := f64(1)
+explicit_harmonics :: proc(harmonics: []f32, volume_limiter: f32, phase: f32) -> f32 {
+    val: f32
+    i := f32(1)
     for n in harmonics {
         val += i * (1.0 / n) * math.sin(2.0 * math.PI * phase * n)
     }
@@ -15,6 +15,6 @@ explicit_harmonics :: proc(harmonics: []f64, volume_limiter: f64, phase: f64) ->
     return val
 }
 
-test :: proc(phase: f64) -> f64 {
+test :: proc(phase: f32) -> f32 {
     return 0
 }
